- add bool and unsigned int to the allowed types - fix one case where floats were used (the JIT always calls with doubles)

+++ src/g_statusbar/sbarinfo.cpp
@@ -1607,7 +1607,7 @@ DEFINE_ACTION_FUNCTION_NATIVE(DSBarInfo, ShowPop, SBarInfo_ShowPop)
 	return 0;
 }
 
-static int SBarInfo_GetProtrusion(DSBarInfo *self, float scale)
+static int SBarInfo_GetProtrusion(DSBarInfo *self, double scale)
 {
 	return self->_GetProtrusion(scale);
 }
